🚚 rename:    대표 인덱스 페이지 외부 모듈 파일 수정
