+++ index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>팀원 소개 및 프로젝트</title>
     <!-- Vite가 번들링한 main.ts 모듈 로드 -->
-    <script type="module" src="/src/main.ts"></script>
+    <script type="module" src="/src/assets/js/introduce.ts"></script>
   </head>
   <body class="bg-gray-100 p-6 space-y-12">
     <!-- 1. 모바일 전용 팀원 리스트 (md 이상 화면에서 숨김) -->
